Move s3 delete file test to `StateSubmissionForm.test.ts`
